requirements.txt
docopt==0.6.2
msgpack-python==0.5.6
paho-mqtt==1.4.0
pika==0.13.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
redis==2.10.5
stomp.py==4.1.21